v0.63 fixes a few bugs that have been discovered since v0.62. This is in parallel to v0.70 development effort that is progressing on the trunk, since that code is not yet ready to be released.
-- Fixed some compiling errors with some uses of SmallMatrix.
-- Disabled the cases where BlasRowMajor was used, since I had problems
   with it on a particular implementation, and I'm not sure if the error
   is in my code or in the BLAS code.  So, I removed these pathways
   for now until I can figure out what might be going on with it.
-- Fixed the return type of some methods that erroneously returned CStyle
   views when the source object was FortranStyle.

There are also two new feature additions.

First, we added a way to access the version of TMV being used.
You can do this either in C++ code with the function TMV_Version().
This returns a string that gives the version being used within your code.
There is also an executable, tmv-version, that can be used to access
the TMV version that is installed at the moment.

Second, we added support for ups, which is probably completely irrelevant 
to the majority of users.  But if you use ups for version control, you can 
use the command scons install WITH_UPS=true to configure tmv in ups.

Also, in response to feedback from a number of TMV users, I decided to 
change the names of many of the methods to start with a lowercase.
The free functions all still start with a capital letter.
Both use camelCase, with underscores after 2 or 3 letter abbreviations
such as LU or SV to help them stand out (since the camel case doesn't do it).
This seems to be a more common standard in the C++ community and is more
self-consistent than the style I had been using.

The old names are still here for this version, but are marked as deprecated.
Compiling old code with g++ with the flag -Wno-deprecated-declarations 
will make is compile ok.  But the deprecated versions will disappear 
eventually, so you will want to covert your existing code over to the new 
names at some point.